[Repo Restructure] Moved All Kivy Components to Specified GUI Framework Folder
This was done because we are planning to revert our decision to move back to PyQt5. Due to technical difficulties, we talked and observed that it is not possible to do dynamic creation of IDS and takes more time than we expected. We also conclude that loading large sets of data from database to UI takes about approx >10 seconds to before normalizing the UI back to it's good condition / performance. @CodexLink