[MAJOR PROGRESSION] > File Clean-up and Various Found Potential Bug Fix
I was now able to see bugs that can be fixed soon enough. For now, I'll commit this one. And keep in mind that this includes broken direction of class due to object name changes and reference name changes (file name). This commit also removes all old files (old files that were known as last-save state) as we should not try to base from there anymore due to possibly work-space issue. Also and lastly, it includes major repository restructure by major renaming of multiple folders. That would be for now.