[Project Progression] Added Passing Back-and-Forth Staff Information
This commit basically says I can give the user information that is passed by LoginCore to WindowController to one another Window. When the Child of WIndowController closes, WindowController can be spawned once again but this time, includes the recently passed arguments that spawns the children of WindowController.

TL;DR > When closing, gives the info back and forth until termination.