[Project Progression] Dynamic State Mechanism of Buttons, Now Working
That was fun to implement :) What's left in this particular system is Add and Edit Functionality. And Database Interchange During Table Change. It may be a few but we are doing something more than what I just listed here. I spent 1 and a half days for me to do this shit. It's a little bit worth of progress in terms of working on this one. But I don't wanna think of playing at the moment because of projects and finals period. Things get pressured but we have to do of what we can do. (We, some parts, yes). Continuous Commit was now finished at the moment. All was done in 1 and a half fucking day. (Sorry for bad words).