[Project Progression] Lost Of Tracking, Phase #1
I'm too scared that my progress might be gone in a minute that I may attempt to do something stupid. Here in this case I modified alot and added new files for this commit. I cannot dice it due to the fact that I'm in a hurry and want to do something for other academic assignments. Anyway, this commit contains A LOT of changes some of it is still progressing. Feel free to git diff this as I dont have time to say or something about the changes. But I know most of the changes though so. Fingers Crossed.

This was a checkpoint phase #1. Stay tuned.